requirements.txt
apache-airflow
confluent-kafka
requests
kafka-python
six